# Conflicts:
#	app/src/main/java/ru/practicum/android/diploma/search/presentation/SearchVacanciesViewModel.kt
#	app/src/main/java/ru/practicum/android/diploma/search/ui/SearchFragment.kt

+++ app/src/main/java/ru/practicum/android/diploma/search/data/dto/reponse/VacanciesResponse.kt
@@ -4,9 +4,9 @@ import com.google.gson.annotations.SerializedName
 
 data class VacanciesResponse(
     val items: List<ItemDto>,
-    val found: Long,
-    val page: Long,
-    val pages: Long,
+    val found: Int,
+    val page: Int,
+    val pages: Int,
     @SerializedName("per_page")
-    val perPage: Long
+    val perPage: Int
 ) : Response()
